feat: Added navigation bar
