+++ src/components/common/state-nav.module.scss
@@ -1,0 +1,73 @@
+@import '~scss/colors.module.scss';
+@import '~scss/breakpoints.module.scss';
+
+.state-nav {
+  position: relative;
+  @media (min-width: $viewport-ms) {
+    background: white;
+    position: sticky;
+    z-index: 1;
+    top: 0;
+    width: calc(
+      100% + 1px
+    ); // this hides the table borders that otherwise appear on the top right
+  }
+  @media (min-width: $viewport-ms) and (max-width: $viewport-md) {
+    width: calc(100vw - 1.5rem);
+    padding-right: 1.5rem;
+  }
+  h2 {
+    margin: 0;
+  }
+  .state-nav-inner {
+    border-bottom: 2px solid $color-slate-500;
+    margin-bottom: 1.5rem;
+    @media (min-width: $viewport-lg) {
+      display: flex;
+      flex-wrap: wrap;
+      width: 100%;
+      > * {
+        flex: 1 1 auto;
+      }
+    }
+  }
+}
+
+.state-nav-combobox {
+  @media (min-width: $viewport-md) {
+    text-align: right;
+  }
+  @media (max-width: $viewport-ms) {
+    padding: 0.75rem 0 0.9rem;
+    display: block;
+  }
+  label {
+    font-weight: bold;
+    display: inline-block;
+  }
+
+  [data-reach-combobox] {
+    margin: 0.3rem 0;
+    @media (min-width: $viewport-ms) {
+      display: inline-block;
+      margin-left: 1rem;
+    }
+  }
+  input {
+    padding: 0.3rem;
+    border: 1px solid $color-slate-600;
+    @media (max-width: $viewport-ms) {
+      width: 100%;
+      padding: 0.5rem 0.4rem;
+    }
+  }
+}
+
+.popover {
+  [data-user-value] {
+    font-weight: bold;
+  }
+  [data-suggested-value] {
+    font-weight: normal;
+  }
+}
