--- src/components/pages/data/state-nav.module.scss
@@ -1,41 +1,0 @@
-@import '../../../scss/colors.module.scss';
-@import '../../../scss/breakpoints.module.scss';
-
-.state-nav-combobox {
-  @media (min-width: $viewport-md) {
-    text-align: right;
-  }
-  @media (max-width: $viewport-ms) {
-    padding: 0.75rem 0 0.9rem;
-    display: block;
-  }
-  label {
-    font-weight: bold;
-    display: inline-block;
-  }
-
-  [data-reach-combobox] {
-    margin: 0.3rem 0;
-    @media (min-width: $viewport-ms) {
-      display: inline-block;
-      margin-left: 1rem;
-    }
-  }
-  input {
-    padding: 0.3rem;
-    border: 1px solid $color-slate-600;
-    @media (max-width: $viewport-ms) {
-      width: 100%;
-      padding: 0.5rem 0.4rem;
-    }
-  }
-}
-
-.popover {
-  [data-user-value] {
-    font-weight: bold;
-  }
-  [data-suggested-value] {
-    font-weight: normal;
-  }
-}
