+++ src/components/pages/race/dashboard/states.module.scss
@@ -1,0 +1,9 @@
+@import '~scss/breakpoints.module.scss';
+
+.header {
+  @media (min-width: $viewport-md) {
+    margin-top: -50px;
+    padding-top: 50px;
+    display: inline-block;
+  }
+}
