+++ src/pages/data/index.js
@@ -1,51 +1,52 @@
 import React from 'react'
 import { graphql } from 'gatsby'
-import DetailText from '../../components/common/detail-text'
-import Layout from '../../components/layout'
-import StateList from '../../components/pages/data/state-list'
-import StatesNoScriptNav from '../../components/pages/data/state-nav-no-script'
-import StatesNav from '../../components/pages/data/state-nav'
-import SummaryTable from '../../components/common/summary-table'
-import { SyncInfobox } from '../../components/common/infobox'
-import stateNavStyles from './index.module.scss'
+import DetailText from '~components/common/detail-text'
+import Layout from '~components/layout'
+import StateList from '~components/pages/data/state-list'
+import StatesNoScriptNav from '~components/pages/data/state-nav-no-script'
+import StatesNav from '~components/common/state-nav'
+import SummaryTable from '~components/common/summary-table'
+import { SyncInfobox } from '~components/common/infobox'
 
-export default ({ data }) => (
-  <Layout
-    title="Most recent data"
-    description="Our most up-to-date data on COVID-19 in the US."
-    navigation={data.allContentfulNavigationGroup.edges[0].node.pages}
-  >
-    <div
-      className="module-content"
-      dangerouslySetInnerHTML={{
-        __html: data.dataPreamble.nodes[0].content.childMarkdownRemark.html,
-      }}
-    />
-    <SyncInfobox />
-    <SummaryTable data={data.allCovidUs.edges[0].node} showOutcomes={false} />
-    <DetailText>
-      <span
+export default ({ data }) => {
+  const stateNavList = []
+  data.allCovidStateInfo.edges.forEach(({ node }) => {
+    stateNavList.push(node)
+  })
+  return (
+    <Layout
+      title="Most recent data"
+      description="Our most up-to-date data on COVID-19 in the US."
+      navigation={data.allContentfulNavigationGroup.edges[0].node.pages}
+    >
+      <div
         className="module-content"
         dangerouslySetInnerHTML={{
-          __html:
-            data.dataSummaryFootnote.nodes[0].content.childMarkdownRemark.html,
+          __html: data.dataPreamble.nodes[0].content.childMarkdownRemark.html,
         }}
       />
-    </DetailText>
-    <div className={`state-nav-header ${stateNavStyles.stateNav}`}>
-      <div className={stateNavStyles.stateNavInner}>
-        <h2 id="states-top">Totals by state</h2>
+      <SyncInfobox />
+      <SummaryTable data={data.allCovidUs.edges[0].node} showOutcomes={false} />
+      <DetailText>
+        <span
+          className="module-content"
+          dangerouslySetInnerHTML={{
+            __html:
+              data.dataSummaryFootnote.nodes[0].content.childMarkdownRemark
+                .html,
+          }}
+        />
+      </DetailText>
+      <StatesNav title="Totals by state" stateList={stateNavList} />
 
-        <StatesNav stateList={data.allCovidStateInfo.edges} />
-      </div>
-    </div>
-    <StatesNoScriptNav stateList={data.allCovidStateInfo.edges} />
-    <StateList
-      states={data.allCovidStateInfo.edges}
-      stateData={data.allCovidState.edges}
-    />
-  </Layout>
-)
+      <StatesNoScriptNav stateList={data.allCovidStateInfo.edges} />
+      <StateList
+        states={data.allCovidStateInfo.edges}
+        stateData={data.allCovidState.edges}
+      />
+    </Layout>
+  )
+}
 
 export const query = graphql`
   query {
