+++ src/pages/race/dashboard.js
@@ -1,34 +1,47 @@
 import React from 'react'
 import { graphql } from 'gatsby'
 import Layout from '../../components/layout'
-
+import StateNav from '~components/common/state-nav'
 import Hero from '~components/pages/race/dashboard/hero'
 import HeaderSorter from '~components/pages/race/dashboard/header-sorter'
 import NoData from '~components/pages/race/dashboard/no-data'
 import States from '~components/pages/race/dashboard/states'
 import UsOverview from '~components/pages/race/dashboard/us-overview'
 
-export default ({ data }) => (
-  <Layout
-    title="Racial Data Dashboard"
-    description="The COVID-19 pandemic isn’t affecting all communities the same way. The COVID Racial Data Dashboard helps us track this inequity by publishing topline racial data compared with state demographic data."
-  >
-    <Hero
-      ledeContent={
-        data.allContentfulSnippet.edges[0].node
-          .childContentfulSnippetContentTextNode.childMarkdownRemark.html
-      }
-    />
-    <UsOverview
-      statesCasesCount={42}
-      statesDeathsCount={35}
-      statesNotReporting={['Colorado', 'North Dakota', 'South Carolina']}
-    />
-    <NoData stateName="North Dakota" />
-    <HeaderSorter stateName="California" stateReports="race" />
-    <States />
-  </Layout>
-)
+export default ({ data }) => {
+  const stateList = []
+  data.allCovidRaceDataSeparate.nodes.forEach(state => {
+    stateList.push(state)
+  })
+  data.allCovidRaceDataCombined.nodes.forEach(state => {
+    stateList.push(state)
+  })
+  return (
+    <Layout
+      title="Racial Data Dashboard"
+      description="The COVID-19 pandemic isn’t affecting all communities the same way. The COVID Racial Data Dashboard helps us track this inequity by publishing topline racial data compared with state demographic data."
+    >
+      <Hero
+        ledeContent={
+          data.allContentfulSnippet.edges[0].node
+            .childContentfulSnippetContentTextNode.childMarkdownRemark.html
+        }
+      />
+      <UsOverview
+        statesCasesCount={42}
+        statesDeathsCount={35}
+        statesNotReporting={['Colorado', 'North Dakota', 'South Carolina']}
+      />
+      <NoData stateName="North Dakota" />
+      <HeaderSorter stateName="California" stateReports="race" />
+      <StateNav
+        title="Race and Ethnicity Data by State"
+        stateList={stateList.sort((a, b) => (a.name < b.name ? -1 : 1))}
+      />
+      <States />
+    </Layout>
+  )
+}
 
 export const query = graphql`
   query {
@@ -41,6 +54,18 @@ export const query = graphql`
             }
           }
         }
+      }
+    }
+    allCovidRaceDataSeparate {
+      nodes {
+        name
+        state
+      }
+    }
+    allCovidRaceDataCombined {
+      nodes {
+        name
+        state
       }
     }
   }
